account delete done

+++ web/src/controller/AccountController.php
@@ -57,17 +57,37 @@ class AccountController extends Controller
      */
     public function deleteAction($id)
     {
+        /**
+         * @var UserAccountModel this object is used to check user information
+         */
         $user = UserAccountController::getCurrentUser();
+        /** user hasn't logged in */
         if($user == null) {
             return;
         }
+
+        /**
+         * @var BankAccountModel this object is used to check current user's authority to the account
+         */
         $bankAccount = $user->getBankAccountByID($id);
-        if($bankAccount !== null) {
-            $bankAccount->delete();
-            $view = new View('accountClosed');
-            $view->addData('deleted', true);
-            echo $view->addData('accountId', $id)->render();
+
+        /** current user is legit to modify the account*/
+        if($bankAccount !== null ) {
+            if ($bankAccount->getBalance() != 0) {
+                /** account has money left, delete failure*/
+                $view = new View('accountClosed');
+                $view->addData('deleted', false);
+                $view->addData('message', "You still have money left in your account!");
+                echo $view->addData('accountId', $id)->render();
+            } else {
+                /** account deleted successfully */
+                $bankAccount->delete();
+                $view = new View('accountClosed');
+                $view->addData('deleted', true);
+                echo $view->addData('accountId', $id)->render();
+            }
         } else {
+            /**  current user is not legit to modify this account*/
             $view = new View('accountClosed');
             $view->addData('deleted', false);
             echo $view->addData('accountId', $id)->render();
